use viam's z-up coordinate system
