+++ web/frontend/src/components/navigation/components/map.svelte
@@ -3,8 +3,8 @@
 import { onMount, createEventDispatcher } from 'svelte';
 import { Map } from 'maplibre-gl';
 import { type LngLat } from '@/api/navigation';
-import { map } from './stores';
-import { style } from './style';
+import { map } from '../stores';
+import { style } from '../style';
 import ThreeLayer from './layer-three.svelte';
 import Waypoints from './waypoints.svelte';
 import RobotMarker from './robot-marker.svelte';
