+++ web/frontend/src/components/navigation/components/marker.svelte
@@ -8,9 +8,9 @@ export let lng = 0;
 export let lat = 0;
 export let scale = 1;
 export let visible = true;
-export let color: string | undefined;
+export let color: string | null = null;
 
-const marker = new Marker({ scale, color });
+const marker = new Marker({ scale, color: color ?? undefined });
 
 $: {
   marker.setLngLat([lng, lat]);
