+++ web/frontend/src/components/navigation/types.ts
@@ -1,3 +1,10 @@
+export type Mat4 = [
+  number, number, number, number,
+  number, number, number, number,
+  number, number, number, number,
+  number, number, number, number
+]
+
 export interface Translation {
   x: number;
   y: number;
